+++ plugins/admission/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 0.5.3
+## 1.1.0
 * Add an `insightsinfo` function to make Insights information available in rego.
+
+## 1.0.0
+* Bump plugin version
 
 ## 0.5.2
 * Update dependencies