+++ plugins/ci/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
 
-## 0.15.2
+## 1.2.1
+* Updated trivy version to 0.23.0
+* Drop root command
+## 1.2.0
+* Adds auto config. file generation by scanning the repository files
+
+## 1.1.1
+* Fix reading helm `valuesFile` and `fluxFile` when on cloned repo context
+* Fix internal `baseFolder` when not in cloned repo context
+
+## 1.1.0
 * Add an `insightsinfo` function to make Insights information available in rego.
+
+## 1.0.0
+* Update plugin version
 
 ## 0.15.1
 * Run apk update